Update README with Style Addition
I basically tried to copy and paste the extended component provided above, but the map would not appear until I defined the style parameters. I wasted a good amount of time trying to render the map and this issue could totally be avoided for those who prefer to use components.