Repaint clusters only once after markers are changed
The destroy hook of the markers removes the marker from the cluster
object which triggers a repaint for each marker that is destoyed. This
severly degrades performance when a large number of markers is involved.
This is fixed by suppressing the repaint during destruction of the
markers and explicitly calling repaint() in the cluster update hook.

Fixes#178